refactor: move TagColorThemes to color_scheme.dart
