+++ frontend/lib/features/tags/tags_edit_screen.dart
@@ -1,118 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:re_view/core/localization/app_localizations.dart';
+import 'package:re_view/core/theme/color_scheme.dart';
 import 'package:re_view/data/models.dart';
 import 'package:re_view/data/hive_manager.dart';
 import 'package:re_view/shared/widgets.dart';
-
-/// 태그 색상 테마
-///
-/// 태그에 적용할 수 있는 5가지 색상 테마를 제공합니다.
-/// 각 테마는 15개의 색상으로 구성되어 있으며, 태그가 15개를 초과하면 순환 방식으로 색상이 재사용됩니다.
-class TagColorTheme {
-  const TagColorTheme(this.name, this.colors);
-
-  /// 테마 이름 (예: '파스텔', '비비드')
-  final String name;
-
-  /// 테마에 포함된 색상 리스트 (ARGB 형식)
-  final List<int> colors;
-
-  /// 사용 가능한 모든 테마
-  static const List<TagColorTheme> themes = [
-    TagColorTheme('파스텔', [
-      0xFFFFDADA,
-      0xFFFFE4C4,
-      0xFFFFF4B3,
-      0xFFE8F5E9,
-      0xFFB3E5FC,
-      0xFFE1BEE7,
-      0xFFF8BBD0,
-      0xFFFFCCBC,
-      0xFFD1C4E9,
-      0xFFC5E1A5,
-      0xFFFFE082,
-      0xFFFFAB91,
-      0xFFCE93D8,
-      0xFFA5D6A7,
-      0xFFB39DDB,
-    ]),
-    TagColorTheme('비비드', [
-      0xFFFF6B6B,
-      0xFFFFAA33,
-      0xFFFFEB3B,
-      0xFF66BB6A,
-      0xFF42A5F5,
-      0xFF9C27B0,
-      0xFFEC407A,
-      0xFFFF7043,
-      0xFF7E57C2,
-      0xFF9CCC65,
-      0xFFFDD835,
-      0xFFFF8A65,
-      0xFFAB47BC,
-      0xFF81C784,
-      0xFF8E24AA,
-    ]),
-    TagColorTheme('네온', [
-      0xFFFF1744,
-      0xFFFF9100,
-      0xFFFFEA00,
-      0xFF00E676,
-      0xFF00B0FF,
-      0xFFD500F9,
-      0xFFFF4081,
-      0xFFFF6E40,
-      0xFF651FFF,
-      0xFF76FF03,
-      0xFFC6FF00,
-      0xFFFF3D00,
-      0xFFE040FB,
-      0xFF00E5FF,
-      0xFFAA00FF,
-    ]),
-    TagColorTheme('소프트', [
-      0xFFEFDBD5,
-      0xFFF3E5DC,
-      0xFFFFF8DC,
-      0xFFE8F4EA,
-      0xFFE0F2F7,
-      0xFFF3E5F5,
-      0xFFFCE4EC,
-      0xFFFBE9E7,
-      0xFFEDE7F6,
-      0xFFE7EED3,
-      0xFFFFF9C4,
-      0xFFFFE0B2,
-      0xFFF1E1F5,
-      0xFFDCEDC8,
-      0xFFE1BEE7,
-    ]),
-    TagColorTheme('어스톤', [
-      0xFFBCAAA4,
-      0xFFD7CCC8,
-      0xFFE6D7C3,
-      0xFFC5E1A5,
-      0xFFB0BEC5,
-      0xFFCE93D8,
-      0xFFF48FB1,
-      0xFFFFAB91,
-      0xFFB39DDB,
-      0xFFA5D6A7,
-      0xFFDCE775,
-      0xFFFFCC80,
-      0xFFBA68C8,
-      0xFF90CAF9,
-      0xFF9FA8DA,
-    ]),
-  ];
-
-  /// 테마 이름으로 테마 객체 찾기
-  ///
-  /// 해당 이름의 테마가 없으면 첫 번째 테마(파스텔)를 반환합니다.
-  static TagColorTheme getTheme(String name) {
-    return themes.firstWhere((t) => t.name == name, orElse: () => themes[0]);
-  }
-}
 
 /// 태그 수정 화면 (Figma 2-3. Modifying Tags)
 ///
@@ -199,7 +90,7 @@ class _TagsEditScreenState extends State<TagsEditScreen> {
   /// 각 태그는 테마의 색상 배열에서 순환하며 색상을 부여받습니다.
   /// 예: 15개 색상 테마에서 16번째 태그는 첫 번째 색상을 받습니다.
   void _assignColors() {
-    final theme = TagColorTheme.getTheme(_currentTheme);
+    final theme = getTagColorTheme(_currentTheme);
     final newTags = <Tag>[];
 
     for (int i = 0; i < _tags.length; i++) {
@@ -283,7 +174,7 @@ class _TagsEditScreenState extends State<TagsEditScreen> {
             Wrap(
               spacing: 8,
               runSpacing: 8,
-              children: TagColorTheme.themes.map((theme) {
+              children: tagColorThemes.map((TagColorTheme theme) {
                 return ChoiceChip(
                   label: Text(
                     AppLocalizations.of(context).getThemeName(theme.name),
@@ -421,7 +312,7 @@ class _TagsEditScreenState extends State<TagsEditScreen> {
     }
 
     // 현재 테마에서 다음 색상 할당
-    final theme = TagColorTheme.getTheme(_currentTheme);
+    final theme = getTagColorTheme(_currentTheme);
     final colorIndex = _tags.length % theme.colors.length;
 
     setState(() {
